refactor(keyring-internal-api): flatten folders
