+++ packages/keyring-internal-api/src/index.ts
@@ -1,4 +1,7 @@
-export * from './internal';
 export * from './btc';
 export * from './eth';
 export * from './sol';
+export * from './api';
+export * from './events';
+export * from './rpc';
+export * from './types';
