--- packages/keyring-internal-api/src/internal/eth/index.ts
@@ -1,1 +1,0 @@
-export type * from './EthKeyring';
