--- packages/keyring-internal-api/src/internal/index.ts
@@ -1,5 +1,0 @@
-export * from './api';
-export type * from './eth';
-export * from './events';
-export * from './rpc';
-export * from './types';
