+++ packages/keyring-internal-api/src/types.ts
@@ -13,9 +13,9 @@ import { exactOptional, object } from '@metamask/keyring-utils';
 import type { Infer, Struct } from '@metamask/superstruct';
 import { boolean, string, number } from '@metamask/superstruct';
 
-import { BtcP2wpkhAccountStruct } from '../btc/types';
-import { EthEoaAccountStruct, EthErc4337AccountStruct } from '../eth/types';
-import { SolDataAccountStruct } from '../sol/types';
+import { BtcP2wpkhAccountStruct } from './btc/types';
+import { EthEoaAccountStruct, EthErc4337AccountStruct } from './eth/types';
+import { SolDataAccountStruct } from './sol/types';
 
 export type InternalAccountType =
   | EthAccountType
